Finish coding week pages
The week page is now completed. The tasks that aren't due in the next 7
days will not be shown, while the ones that are due will be visible,
even if they were invisible from before. This is to allow the user to
switch between day and week pages without any problems.